svg/animations/repeatcount-numeric-limit.tentative.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><title>A huge &apos;repeatCount&apos; (1e+309) is treated as unspecified</title><h:link href="https://svgwg.org/specs/animations/#TimingAttributes" rel="help"/><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><rect width="50" height="100" fill="#00f"><animate fill="freeze" attributeName="fill" dur="10ms" from="#007f00" repeatCount="Infinity" to="green"/></rect><rect width="50" height="100" x="50" fill="#00f"><animate fill="freeze" attributeName="fill" dur="10ms" from="#007f00" repeatCount="Infinity" to="green"/></rect><script>promise_test(t =&gt; { let watchers = Array.from(document.getElementsByTagName(&apos;animate&apos;)).map(element =&gt; { let watcher = new EventWatcher(t, element, [&apos;endEvent&apos;, &apos;repeatEvent&apos;]); return watcher.wait_for(&apos;endEvent&apos;).then(() =&gt; { assert_equals(getComputedStyle(element).fill, &apos;rgb(0, 128, 0)&apos;); }); }); return Promise.all(watchers); });</script></svg>
